Add pipenv support and improve handling of whitespace in paths
Adds support for pipenv for management of virtual environments. The
old requirements.txt file is maintained for backwards compatibility.
For now, it is necessary to keep the two in sync manually.

In convert.py, directories are now created with os.makedirs rather
than the previous use of os.system. This method is more robust as it
now handles file paths with spaces in them.

When an output file is created that has whitespace in its path, a
warning message is printed to the console as whitespace may interfere
with other applications in the analysis pipeline. The check is done in
util.py but the output is generated in each UI module to help separate
business and UI logic.

An additional test case has been added to check whitespace. At present,
this test copies the existing test MRI data. This is not ideal and a
better implementation could probably be achieved by mocking or breaking
up different elements of the test to a finer level of detail but this
is a little tricky as multiple functions currently use the output file
path including __init__.

This commit also add pytest as a development dependency. This is required
for adding the ability to test for raised exceptions. The use of pytest
over nose is due to lack of ongoing development of nose. See #48.