Update handling of filepaths with whitespace
An warning message is now printed to the console when the output file
path contains whitespace as this may cause problems for other elements
of the analysis pipeline. The testing for whitespace is done as a
general utility function but the output is generated in the different
UI modules to keep business logic separated from UI logic.

The unit testing logic for whitespace handling has been modified. It
now creates a temporary copy of the testing data and tests against that.
This is not an ideal method but at the moment code that uses the
file path is spread between multiple functions in the class including the
__init__ function which makes it difficult to mock.