Add initial support for displaying study metadata
Add the BrukerMetadata class. BrukerMetadata objects represent
metadata relating to a single MRI study including information about
the subject, acquisition parameters and visualisation paramaters.
This enables the display of this metadata to the user to help them
understand what scans have been acquired and to help decide which
scans to convert to nifti.

The bruker2nifti cli has been modified to now accept a command as the
first positional parameter when running the application. If 'convert'
is specified, then the study will be converted as is currently the case.
If 'list' is supplied, this should list all scans and reconstructions
in the study. For now, the implementation of list is a minimal stub.
For backwards compatibility, if no command is specified, it defaults to
convert and functions just as how the cli is currently run.

Included in this commit is a small fix to the bruker_read_files function
to properly read the specified reconstruction metadata rather than
always converting reconstruction '1'.

Pytest test fixtures have been incorporated into the test framework. New
test fixtures should be specified in 'test/fixtures' and imported into
test/conftest.py.