Finish simple implementation of study listing
Running the cli with `bruker2nifti list -i ...` will now print a
brief summary of the study including a the scans that comprise it
with some basic information. This can be helpful in quickly getting
an overview of what scans are available and help plan selective
conversion to the nifti format for further processing/analysis.