Fix specification of scans to convert in bruker2nii
The current implementation of the -scans_list option to bruker2nii
only works for single-digit scan numbers. This adds a new --scans
option that works with multi-digit scan numbers. Scan numbers should
be separated by whitespace or commas. If both --scans and -scan_list
is specified, --scans will take priority. -scan_list is kept
unchanged for backwards compatibility.

Other UIs (bruker2nii_scan and the tkinter GUI are not altered).

The existing implementation of the Bruker2Nifti class doesn't work
well with specifying a list of scans to convert because output
filenames are derived during object instantiation whilst specific
scans to convert must be specified after instantiation. A workaround
is used in this commit though in the future, a better option would
be to refactor the Bruker2Nifti class implementation so this isn't
necessary.